Code for figures for poster Eco-Evo GRS/GRC 2025
Fixed an issue with the replicates where I only stored the last adaptive walk instead of all the replicates. Still need to fix the same issue with the sets of landscapes I generate (I only store the last set of landscapes I generate for the last combination of IM_k and IM_c).
I added the code to generate all the figures from my poster for the GRS/GRC in July 2025.
Updated the date in in the ReadMe file